[PATCH] Cleanup and switch to pyproject.toml
