--- requirements.txt
@@ -1,3 +1,0 @@
-requests~=2.32.3
-typer~=0.12.5
-rich~=13.9.3
